requirements.txt
watchdog==2.0.2
inotify==0.2.10
pyyaml
pandas
sklearn
matplotlib
pyspark